+++ tests/test_thermal.py
@@ -24,20 +24,20 @@ import felupe as fem
 def test_thermal():
     mesh = fem.Rectangle(n=3)
     region = fem.RegionQuad(mesh)
-    temperature = fem.Field(region, dim=1)
+    temperature = fem.Field(region, dim=1, values=293.15)
     field = fem.FieldContainer([temperature])
 
-    region_heat_transfer = fem.RegionQuadBoundary(mesh, mask=mesh.y == 1.0)
-    temperature_heat_transfer = fem.Field(region_heat_transfer, dim=1)
-    field_heat_transfer = fem.FieldContainer([temperature_heat_transfer])
+    region_bottom = fem.RegionQuadBoundary(mesh, mask=mesh.y == 0.0)
+    temperature_bottom = fem.Field(region_bottom, dim=1)
+    field_bottom = fem.FieldContainer([temperature_bottom])
 
-    region_flux = fem.RegionQuadBoundary(mesh, mask=mesh.y == 1.0)
-    temperature_flux = fem.Field(region_flux, dim=1)
-    field_flux = fem.FieldContainer([temperature_flux])
+    region_top = fem.RegionQuadBoundary(mesh, mask=mesh.y == 1.0)
+    temperature_top = fem.Field(region_top, dim=1)
+    field_top = fem.FieldContainer([temperature_top])
 
     boundaries = fem.BoundaryDict(
-        left=fem.Boundary(temperature, fx=0),
-        right=fem.Boundary(temperature, fx=1),
+        left=fem.Boundary(temperature, fx=0, value=293.15),
+        right=fem.Boundary(temperature, fx=1, value=293.15),
     )
 
     solid = fem.thermal.SolidBodyThermal(
@@ -58,13 +58,19 @@ def test_thermal():
     )
 
     heat_transfer = fem.thermal.SolidBodySurfaceHeatTransfer(
-        field=field_heat_transfer,
+        field=field_top,
         coefficient=1.0,  # W/(m^2*K)
-        temperature=10.0,  # K
+        temperature=293.15 + 10.0,  # K
+    )
+
+    heat_radiation = fem.thermal.SolidBodySurfaceRadiation(
+        field=field_top,
+        emissivity=0.8,  # dimensionless, between 0 and 1
+        temperature=293.15 + 10.0,  # K
     )
 
     heat_flux = fem.thermal.SolidBodyHeatFlux(
-        field=field_flux,
+        field=field_bottom,
         heat_flux=1.0,  # W/m^2
     )
 
@@ -77,14 +83,19 @@ def test_thermal():
 
     time = fem.thermal.TimeStep([solid, heat_transfer, heat_flux])
     table = fem.math.linsteps([0, 0, 1], num=2)
+    table_emissivity = fem.math.linsteps([1, 1, 1], num=2) * 0.8
     ramp = {
         boundaries["right"]: 10 * table,
         time: 0.1 * table,
         heat_transfer: 100 * table,
         heat_flux: 10 * table,
+        heat_radiation: 100 * table,
+        heat_radiation["emissivity"]: table_emissivity,
     }
     step = fem.Step(
-        items=[time, solid, heat_transfer, heat_flux], ramp=ramp, boundaries=boundaries
+        items=[time, solid, heat_transfer, heat_flux, heat_radiation],
+        ramp=ramp,
+        boundaries=boundaries,
     )
     job = fem.Job(steps=[step]).evaluate(
         filename="result.xdmf",  # result file for Paraview
@@ -100,13 +111,13 @@ def test_thermal_axi():
     temperature = fem.Field(region, dim=1)
     field = fem.FieldContainer([temperature])
 
-    region_heat_transfer = fem.RegionQuadBoundary(mesh, mask=mesh.y == 1.0)
-    temperature_heat_transfer = fem.FieldAxisymmetric(region_heat_transfer, dim=1)
-    field_heat_transfer = fem.FieldContainer([temperature_heat_transfer])
+    region_bottom = fem.RegionQuadBoundary(mesh, mask=mesh.y == 0.0)
+    temperature_bottom = fem.Field(region_bottom, dim=1)
+    field_bottom = fem.FieldContainer([temperature_bottom])
 
-    region_flux = fem.RegionQuadBoundary(mesh, mask=mesh.y == 1.0)
-    temperature_flux = fem.FieldAxisymmetric(region_flux, dim=1)
-    field_flux = fem.FieldContainer([temperature_flux])
+    region_top = fem.RegionQuadBoundary(mesh, mask=mesh.y == 1.0)
+    temperature_top = fem.Field(region_top, dim=1)
+    field_top = fem.FieldContainer([temperature_top])
 
     boundaries = fem.BoundaryDict(
         left=fem.Boundary(temperature, fx=0),
@@ -131,13 +142,13 @@ def test_thermal_axi():
     )
 
     heat_transfer = fem.thermal.SolidBodySurfaceHeatTransfer(
-        field=field_heat_transfer,
+        field=field_top,
         coefficient=1.0,  # W/(m^2*K)
         temperature=10.0,  # K
     )
 
     heat_flux = fem.thermal.SolidBodyHeatFlux(
-        field=field_flux,
+        field=field_bottom,
         heat_flux=1.0,  # W/m^2
     )