+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file. The format 
 
 ### Added
 - Add Becker's logarithmic material model.
-- Add `felupe.thermal` with `thermal.SolidBodyThermal`, `thermal.SolidBodySurfaceHeatTransfer`, `thermal.SolidBodyHeatFlux` and `thermal.TimeStep`. The time step class must include all thermal items and must be the first item in a step. Thermal transient zero-valued time-steps are supported. The stationary solution is obtained by `time_step=None`, whereas `time_step=0.0` is treated as a time-increment of length zero.
+- Add `felupe.thermal` with `thermal.SolidBodyThermal`, `thermal.SolidBodySurfaceHeatTransfer`, `thermal.SolidBodySurfaceRadiation`, `thermal.SolidBodyHeatFlux` and `thermal.TimeStep`. The time step class must include all thermal items and must be the first item in a step. Thermal transient zero-valued time-steps are supported. The stationary solution is obtained by `time_step=None`, whereas `time_step=0.0` is treated as a time-increment of length zero.
 - Add `Mesh.cells_in(other_mesh, decimals=8)` to generate a cells-mask for cells, which are also included in another mesh.
 - Add keyword-based ramps for items: `mechanics.UpdateItem` is internally used to access a keyword-specific update method of an item. E.g., for `heat_transfer = thermal.SolidBodySurfaceHeatTransfer(...),` define the heat transfer `coefficient` in a `Step` by `ramp = {heat_transfer["coefficient"]: np.linspace(1.0, 1.1)}`.
 - Add `"temperature"` and `"coefficient"` as keyword-specific update-methods for `thermal.SolidBodySurfaceHeatTransfer`.